# docs/architecture/openapi/speakeasy-api.yaml
openapi: 3.0.3
info:
  title: SpeakEasy API
  version: 0.1.0
  description: 'Machine-readable OpenAPI source of truth for SpeakEasy server-backed Product Base behavior, P0 commercial
    subscription readiness, and P0.1 expression automation training. P0.2/P1/P2 roadmap items are intentionally
    excluded from implementation-level paths until they have owning increments.

    '
  license:
    name: Proprietary
servers:
- url: /v1
  description: Versioned API base path
security:
- bearerAuth: []
...
paths:
...
  /ai/transcribe:
    post:
      tags:
      - AI Gateway
      operationId: transcribeAudio
      summary: Transcribe uploaded or referenced learner audio.
      x-traceability:
        product_objects:
        - product-base
        - p0-1-expression-automation-training
        - p0-commercial-readiness
        requirements:
        - FR-004
        - FR-007
        - P01-FR-006
        - FR-COM-010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranscribeRequest'
            examples:
              default:
                $ref: '#/components/examples/TranscribeRequestExample'
      responses:
        '200':
          description: Transcription result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeResponse'
              examples:
                default:
                  $ref: '#/components/examples/TranscribeResponseExample'
        '429':
          $ref: '#/components/responses/UsageLimitExceeded'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /ai/tts:
    post:
      tags:
      - AI Gateway
      operationId: synthesizeSpeech
      summary: Synthesize speech from reviewed text.
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-004
        - FR-008
        - FR-COM-010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TtsRequest'
            examples:
              default:
                $ref: '#/components/examples/TtsRequestExample'
      responses:
        '200':
          description: TTS result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TtsResponse'
              examples:
                default:
                  $ref: '#/components/examples/TtsResponseExample'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  $ref: '#/components/examples/ErrorResponseExample'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /ai/coach-turn:
    post:
      tags:
      - AI Gateway
      operationId: coachTurn
      summary: Generate structured coach feedback candidate.
      x-traceability:
        product_objects:
        - product-base
        - p0-1-expression-automation-training
        - p0-commercial-readiness
        requirements:
        - FR-008
        - P01-FR-007
        - FR-COM-010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachTurnRequest'
            examples:
              default:
                $ref: '#/components/examples/CoachTurnRequestExample'
      responses:
        '200':
          description: Structured feedback candidate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachTurnResponse'
              examples:
                default:
                  $ref: '#/components/examples/CoachTurnResponseExample'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  $ref: '#/components/examples/ErrorResponseExample'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /ai/feedback:
    post:
      tags:
      - AI Gateway
      operationId: generateFeedback
      summary: Generate structured practice feedback candidate.
      x-traceability:
        product_objects:
        - product-base
        - p0-1-expression-automation-training
        - p0-commercial-readiness
        requirements:
        - FR-008
        - P01-FR-007
        - FR-COM-010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
            examples:
              default:
                $ref: '#/components/examples/FeedbackRequestExample'
      responses:
        '200':
          description: Feedback candidate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachTurnResponse'
              examples:
                default:
                  $ref: '#/components/examples/CoachTurnResponseExample'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  $ref: '#/components/examples/ErrorResponseExample'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /ai/pronunciation:
    post:
      tags:
      - AI Gateway
      operationId: scorePronunciation
      summary: Produce pronunciation or completeness score signal.
      x-traceability:
        product_objects:
        - product-base
        - p0-1-expression-automation-training
        - p0-commercial-readiness
        requirements:
        - FR-004
        - FR-008
        - P01-FR-007
        - FR-COM-010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PronunciationRequest'
            examples:
              default:
                $ref: '#/components/examples/PronunciationRequestExample'
      responses:
        '200':
          description: Score signal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PronunciationResponse'
              examples:
                default:
                  $ref: '#/components/examples/PronunciationResponseExample'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  $ref: '#/components/examples/ErrorResponseExample'
    parameters:
    - $ref: '#/components/parameters/RequestId'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    opsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Restricted operations token.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Provider-Signature
  parameters:
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
...
  responses:
...
    UsageLimitExceeded:
      description: Usage limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
...
  schemas:
    SchemaVersion:
      type: integer
      enum:
      - 1
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorCode:
      type: string
      enum:
      - UNAUTHENTICATED
      - FORBIDDEN
      - ENTITLEMENT_REQUIRED
      - USAGE_LIMIT_EXCEEDED
      - INVALID_RECEIPT
      - PRODUCT_MISMATCH
      - SUBSCRIPTION_EXPIRED
      - IDEMPOTENCY_CONFLICT
      - SCHEMA_VALIDATION_FAILED
      - PROVIDER_UNAVAILABLE
      - DELETE_IN_PROGRESS
      - RESOURCE_NOT_FOUND
      - CONFLICT
...
    CoachFeedback:
      type: object
      required:
      - summary
      properties:
        summary:
          type: string
        feedback_type:
          type: string
          enum:
          - next_question
          - retry
          - expression_suggestion
          - score_signal
          - recoverable_error
        main_issue_type:
          type: string
        suggested_expression:
          type: string
        next_prompt:
          type: string
        score_signal:
          $ref: '#/components/schemas/ScoreSignal'
        validation_status:
          type: string
          enum:
          - valid
          - invalid_schema
          - fallback
        provider_status:
          type: string
          enum:
          - success
          - timeout
          - provider_unavailable
          - invalid_schema
...
    ScoreSignal:
      type: object
      required:
      - score_kind
      - status
      - source
      properties:
        score_kind:
          type: string
          enum:
          - pronunciation
          - completeness
          - grammar
          - task_completion
          - fluency
        value:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        confidence:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        status:
          type: string
          enum:
          - available
          - low_confidence
          - unavailable
        source:
          type: string
          enum:
          - server_side_adapter
...
    TranscribeRequest:
      type: object
      additionalProperties: false
      required:
      - schema_version
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        audio_ref:
          type: string
        language_hint:
          type: string
    TranscribeResponse:
      type: object
      required:
      - schema_version
      - transcript
      - status
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        transcript:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        status:
          type: string
          enum:
          - available
          - no_result
          - provider_unavailable
    TtsRequest:
      type: object
      additionalProperties: false
      required:
      - schema_version
      - text
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        text:
          type: string
        voice:
          type: string
    TtsResponse:
      type: object
      required:
      - schema_version
      - audio_ref
      - status
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        audio_ref:
          type: string
        status:
          type: string
          enum:
          - available
          - provider_unavailable
    CoachTurnRequest:
      type: object
      additionalProperties: false
      required:
      - schema_version
      - session_id
      - transcript
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        session_id:
          type: string
        transcript:
          type: string
        target_expression_ids:
          type: array
          items:
            type: string
    FeedbackRequest:
      allOf:
      - $ref: '#/components/schemas/CoachTurnRequest'
    CoachTurnResponse:
      type: object
      required:
      - schema_version
      - feedback
      - validation_status
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        feedback:
          $ref: '#/components/schemas/CoachFeedback'
        validation_status:
          type: string
          enum:
          - valid
          - invalid_schema
          - fallback
    PronunciationRequest:
      type: object
      additionalProperties: false
      required:
      - schema_version
      - audio_ref
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        audio_ref:
          type: string
        reference_text:
          type: string
    PronunciationResponse:
      type: object
      required:
      - schema_version
      - score_signal
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        score_signal:
          $ref: '#/components/schemas/ScoreSignal'
...
  examples:
    ErrorResponseExample:
      summary: Standard error response
      value:
        error:
          code: SCHEMA_VALIDATION_FAILED
          message: Request validation failed.
          request_id: req_20260526_sample
          details:
            field: schema_version
            reason: Unsupported schema version.
...
    TranscribeRequestExample:
      summary: transcribeAudio request example
      value:
        schema_version: 1
        audio_ref: sample
        language_hint: en-US
    TranscribeResponseExample:
      summary: transcribeAudio response 200 example
      value:
        schema_version: 1
        transcript: Sample text.
        confidence: 0.85
        status: available
    TtsRequestExample:
      summary: synthesizeSpeech request example
      value:
        schema_version: 1
        text: Sample text.
        voice: sample
    TtsResponseExample:
      summary: synthesizeSpeech response 200 example
      value:
        schema_version: 1
        audio_ref: sample
        status: available
    CoachTurnRequestExample:
      summary: coachTurn request example
      value:
        schema_version: 1
        session_id: session_id_sample
        transcript: Sample text.
        target_expression_ids:
        - target_expression_ids-sample
    CoachTurnResponseExample:
      summary: coachTurn response 200 example
      value:
        schema_version: 1
        feedback:
          summary: sample
          feedback_type: next_question
          main_issue_type: sample
          suggested_expression: sample
          next_prompt: sample
          validation_status: valid
          provider_status: success
          score_signal:
            score_kind: pronunciation
            value: 0.85
            confidence: 0.85
            status: available
            source: server_side_adapter
        validation_status: valid
    FeedbackRequestExample:
      summary: generateFeedback request example
      value:
        schema_version: 1
        session_id: session_id_sample
        transcript: Sample text.
        target_expression_ids:
        - target_expression_ids-sample
    PronunciationRequestExample:
      summary: scorePronunciation request example
      value:
        schema_version: 1
        audio_ref: sample
        reference_text: Sample text.
    PronunciationResponseExample:
      summary: scorePronunciation response 200 example
      value:
        schema_version: 1
        score_signal:
          score_kind: pronunciation
          value: 0.85
          confidence: 0.85
          status: available
          source: server_side_adapter
...